move DefaultLayout.tsx to component dir
