+++ frontend/src/components/DefaultLayout.tsx
@@ -1,6 +1,6 @@
 import { ChakraProvider, Box, Flex, theme } from "@chakra-ui/react";
-import { TopHeader } from "../components/TopHeader";
-import { SideMenu } from "../components/SideMenu";
+import { TopHeader } from "./TopHeader";
+import { SideMenu } from "./SideMenu";
 import React, { ReactNode } from 'react';
 
 interface DefaultLayoutProps {
