+++ frontend/src/routes/Account.tsx
@@ -1,4 +1,4 @@
-import { DefaultLayout } from "./DefaultLayout";
+import { DefaultLayout } from "../components/DefaultLayout";
 
 export const Account = () => <DefaultLayout><h1>アカウント関連、個別の課題状況</h1></DefaultLayout>;
