+++ frontend/src/routes/Message.tsx
@@ -1,4 +1,4 @@
-import { DefaultLayout } from "./DefaultLayout";
+import { DefaultLayout } from "../components/DefaultLayout";
 
 export const Message = () => <DefaultLayout><h1>課題のopen,closeの情報の表示</h1></DefaultLayout>;
