+++ frontend/src/routes/Schedule.tsx
@@ -1,4 +1,4 @@
-import { DefaultLayout } from "./DefaultLayout";
+import { DefaultLayout } from "../components/DefaultLayout";
 
 export const Schedule = () => <DefaultLayout><h1>締め切りとかのカレンダー的なものとか</h1></DefaultLayout>;
